+++ addon/components/groupedit-toolbar.js
@@ -61,6 +61,15 @@ export default FlexberryBaseComponent.extend({
   */
   deleteButton: true,
 
+  /**
+    Flag: indicates whether to show default settings button at toolbar.
+
+    @property defaultSettingsButton
+    @type Boolean
+    @default true
+  */
+  defaultSettingsButton: true,
+
   actions: {
     /**
       Handles add record button click and triggers add record event on
@@ -98,6 +107,32 @@ export default FlexberryBaseComponent.extend({
 
       let componentName = this.get('componentName');
       this.get('_groupEditEventsService').deleteRowsTrigger(componentName);
+    },
+
+    /**
+      Handles default usersettings button click.
+
+      @method actions.setDefaultSettings
+    */
+    setDefaultSettings() {
+      let componentName = this.get('componentName');
+      let userSettingsService = this.get('userSettingsService');
+      let _this = this;
+
+      if (!userSettingsService.haveDefaultUserSetting(componentName)) {
+        return;
+      }
+
+      let defaultDeveloperUserSetting = userSettingsService.getDefaultDeveloperUserSetting(componentName) || {};
+      let currentUserSetting = userSettingsService.getCurrentUserSetting(componentName);
+      currentUserSetting.sorting = defaultDeveloperUserSetting.sorting || [];
+      currentUserSetting.colsOrder = defaultDeveloperUserSetting.colsOrder;
+      currentUserSetting.columnWidths = defaultDeveloperUserSetting.columnWidths;
+      userSettingsService.saveUserSetting(componentName, undefined, currentUserSetting)
+      .then(record => {
+        this.set('sorting', currentUserSetting.sorting);
+        _this.get('_groupEditEventsService').updateWidthTrigger(componentName);
+      });
     }
   },